refactor(api): extracted api work to a separate package
