+++ qrimatic/scheduler/deploy_test.go
@@ -1,4 +1,4 @@
-package update
+package scheduler
 
 import (
 	"context"
